requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
hupper==1.12
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
passlib==1.7.4
PasteDeploy==3.0.1
plaster==1.1.2
plaster-pastedeploy==1.0.1
psycopg==3.1.13
pyasn1==0.5.1
pycparser==2.21
pydantic==1.10.13
pydantic_core==2.14.5
pyramid==2.0.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
setuptools==68.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlmodel==0.0.12
starlette==0.27.0
translationstring==1.4
typing_extensions==4.8.0
uvicorn==0.24.0.post1
venusian==3.1.0
WebOb==1.8.7
wheel==0.41.2
zope.deprecation==5.0
zope.interface==6.1
